third_party/WebKit/LayoutTests/http/tests/media/video-load-metadata-decode-error-expected.txt

This test case simulates a decode error after loading meta data of a video.

EVENT(loadedmetadata)
loaded metadata of media file
EVENT(error)
failed to load media file
EXPECTED (video.networkState == '1') OK
END OF TEST
